Removed Sheets systems, continued updating tasks for v5
