+++ lists.py
@@ -1,19 +1,17 @@
-from src.sheets.events import get_events
-
 async def get_state_list():
-    return ["Alabama (AL)", 
-    "Alaska (AK)", 
-    "Arizona (AZ)", 
-    "Arkansas (AR)", 
-    "California (North) (CA-N)", 
-    "California (South) (CA-S)", 
-    "Colorado (CO)", 
+    return ["Alabama (AL)",
+    "Alaska (AK)",
+    "Arizona (AZ)",
+    "Arkansas (AR)",
+    "California (North) (CA-N)",
+    "California (South) (CA-S)",
+    "Colorado (CO)",
     "Connecticut (CT)",
     "Delaware (DE)",
     "District of Columbia (DC)",
     "Florida (FL)",
     "Georgia (GA)",
-    "Hawaii (HI)", 
+    "Hawaii (HI)",
     "Idaho (ID)",
     "Illinois (IL)",
     "Indiana (IN)",
@@ -34,7 +32,7 @@ async def get_state_list():
     "New Hampshire (NH)",
     "New Jersey (NJ)",
     "New Mexico (NM)",
-    "New York (NY)", 
+    "New York (NY)",
     "North Carolina (NC)",
     "North Dakota (ND)",
     "Ohio (OH)",
