+++ src/mongo/mongo.py
@@ -46,6 +46,9 @@ async def get_tags():
 async def get_reports():
     return await get_entire_collection("data", "reports")
 
+async def get_events():
+    return await get_entire_collection("data", "events")
+
 async def insert(db_name, collection_name, insert_dict):
     global client
     collection = client[db_name][collection_name]
