--- src/sheets/__init__.py
@@ -1,3 +1,0 @@
-"""
-USED FOR CONNECTING TO GOOGLE SHEETS AND GETTING/PUTTING INFORMATION
-"""
